requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-community==0.0.27
requests==2.31.0
beautifulsoup4==4.12.3
plotly==5.18.0
python-dotenv==1.0.1
protobuf==3.20.3
